+++ .eslintrc.js
@@ -11,4 +11,12 @@ module.exports = {
   parser: '@typescript-eslint/parser',
   plugins: ['@typescript-eslint'],
   extends: ['plugin:@typescript-eslint/eslint-recommended', 'plugin:prettier/recommended'],
+  rules: {
+    'prettier/prettier': [
+      'error',
+      {
+        endOfLine: 'auto',
+      },
+    ],
+  },
 };
