+++ .gitignore
@@ -4,3 +4,5 @@
 .idea
 .vscode
 .DS_Store
+
+test.*
