+++ src/index.ts
@@ -1,8 +1,8 @@
 export { connect } from './connection';
-export type { Knex } from './connection';
+export type { Knex, Mongoose, MongoConfig } from './connection';
 export type { IdType, RowType } from './crud.type';
 export { CrudOperations } from './crud-operations';
-export type { CrudFilter, CrudFilterColumns, CrudOperationsOpts } from './crud-operations';
+export type { CrudFilter, CrudFilterColumns, CrudOperationsOpts } from './crud-operations.interface';
 export { SortOrder, sort, sortAsc, sortDesc, parseSort, parseSorts } from './sort';
 export type { Sort } from './sort';
 export { relation, parseRelation, parseRelations } from './relation';
@@ -12,7 +12,5 @@ export type { WeaverOpts } from './weaver';
 export { ExtraOperations } from './extra-operations';
 export type { ExtraOperationsOpts } from './extra-operations';
 
-export { mongoConnect } from './mongo-connection';
-export type { Mongoose } from './mongo-connection';
 export { MongoCrudOperations, Schema } from './mongo-crud-operations';
-export type { MongoCrudOperationsOpts } from './mongo-crud-operations';
+export type { MongoCrudOperationsOpts } from './crud-operations.interface';
